Renaming files and removing redundant functions
