+++ js/phylo.js
@@ -84,74 +84,6 @@ function readTree(text) {
 
 
 /**
- * Recursive function for traversal of tree
- * (output parent before children).
- * @param {object} node
- * @param {string} order: 'preorder' or 'postorder' traversal
- * @param {Array} list: an Array of nodes
- * @return An Array of nodes in pre-order
- */
-function traverse(node, order='preorder', list=Array()) {
-    if (order=='preorder') list.push(node);
-    for (var i=0; i < node.children.length; i++) {
-        list = traverse(node.children[i], order, list);
-    }
-    if (order=='postorder') list.push(node);
-    return(list);
-}
-
-
-/**
- * Convert parsed Newick tree from readTree() into more convenient
- * tabular data frame.
- * @param {object} tree: Return value of readTree
- * @param {boolean} sort: if true, sort data frame by node name
- * @return Array of Objects
- */
-function fortify(tree, sort=true) {
-    var df = [];
-
-    for (const node of traverse(tree, 'preorder')) {
-        if (node.parent === null) {
-            df.push({
-                'parentId': null,
-                'parentLabel': null,
-                'thisId': node.id,
-                'thisLabel': node.label,
-                'children': node.children.map(x=>x.id),
-                'branchLength': 0.,
-                'isTip': (node.children.length===0),
-                'x': node.x,
-                'y': node.y,
-                'angle': node.angle
-            })
-        }
-        else {
-            df.push({
-                'parentId': node.parent.id,
-                'parentLabel': node.parent.label,
-                'thisId': node.id,
-                'thisLabel': node.label,
-                'children': node.children.map(x=>x.id),
-                'branchLength': node.branchLength,
-                'isTip': (node.children.length===0),
-                'x': node.x,
-                'y': node.y,
-                'angle': node.angle
-            })
-        }
-    }
-
-    if (sort) {
-        df = df.sort(function(a, b) {
-            return a.thisId - b.thisId;
-        })
-    }
-    return(df);
-}
-
-
-/**
  * Generate edge list with x,y coordinates extracted from the
  * respective nodes.
  * @param {Array} df:  tabular data frame from fortify()
